Fix Sector3Time truncation in time trial packet parsing (#408)
## Summary

- `Sector3Time` in `PacketToTimeTrialData.ExtractTimeTrialData` was read
as a `byte` instead of a `uint`, per the 2024+ telemetry spec (`uint
m_sector3TimeInMS`). Only the least significant byte of the value ended
up being stored, so real sector times (>= 20,000 ms) were always wrong.
The offset advance of 4 bytes was already correct.
- Changed the read to `Unsafe.ReadUnaligned<uint>`, matching the
existing `Sector1Time`/`Sector2Time` handling.

## Test plan

- Added
`PacketTimeTrial2024Tests.PacketTimeTrialPlayerSessionBestSector3Time2024IsNotTruncatedToByte`,
which parses the existing `F1-2024-TimeTrial.packet` sample and asserts
`Sector3Time` equals the full uint value (20991) instead of the
truncated byte value (255).
- `dotnet build F1Server.Core/F1Server.Core.csproj` — succeeds with no
new warnings.
- `reihitsu-format --check` on both changed files — clean.
- `dotnet test F1Server.Tests/F1Server.Tests.csproj` — all 622 tests
pass.

Closes #76

+++ F1Server.Core/Packets/PacketToObject/PacketToTimeTrialData.cs
@@ -172,7 +172,7 @@ internal class PacketToTimeTrialData(PacketHeader packetHeader) : PacketToXBase(
 
                 actOffset += ConstData.TypeUInt32;
 
-                timeTrialData.Sector3Time = Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
+                timeTrialData.Sector3Time = Unsafe.ReadUnaligned<uint>(ref Unsafe.Add(ref dataPacket, actOffset));
 
                 actOffset += ConstData.TypeUInt32;
